give matcher class its own header and implementation files
